feat(server): add bare bone request handler to server
